% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronabr.R
\docType{package}
\name{coronabr}
\alias{coronabr}
\title{coronabr: Download de dados de COVID-19}
\description{
coronabr coronabr é um pacote de R para fazer download e visualizar os dados
dos casos diários de coronavírus (COVID-19) disponibilizados por diferentes
fontes
}
