% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_corona_br.R
\name{fill_corona_br}
\alias{fill_corona_br}
\title{Preenche as datas vazias nos dados de COVID-19 do Brasil.io}
\usage{
fill_corona_br(df)
}
\arguments{
\item{df}{Data frame contendo o resultado da busca de `get_corona_br()`}
}
\description{
Dado que na ausência da emissão do boletim por alguma secretaria, a data é excluída das planilhas, incluímos a opção de preencher a lacuna com os dados do último boletim. O preenchimento é apenas para os municípios e estados com casos registrados.
}
