% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corona.R
\name{get_corona}
\alias{get_corona}
\title{Extrai dados de coronavírus para o Brasil}
\usage{
get_corona(dir = "output", filename = "minsaude", uf = NULL)
}
\arguments{
\item{dir}{Diretório onde salvar o arquivo}

\item{filename}{Nome do arquivo, valor predeterminado "minsaude"}

\item{uf}{Caractere indicando a abreviação do(s) estado(s) brasileiro(s)}
}
\description{
Esta função extrai os valores compilados pelo ministério da Saúde do Brasil
(disponível em: 'https://covid.saude.gov.br/') e salva o resultado no disco.
}
