% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corona_br.R
\name{get_corona_br}
\alias{get_corona_br}
\title{Extrai dados de corona vírus para o Brasil de O Brasil em dados abertos}
\usage{
get_corona_br(dir = "output/", filename = "corona_brasil",
  cidade = NULL, uf = NULL, cidade_ibge_cod = NULL)
}
\arguments{
\item{dir}{Diretório onde salvar o arquivo}

\item{filename}{Nome do arquivo}

\item{cidade}{Caractere indicando o(s) nome(s) do(s) município(s) brasileiro(s). Atenção, fornecer também o estado para evitar ambiguidade}

\item{uf}{Caractere indicando a abreviação do(s) estado(s) brasileiro(s)}

\item{cidade_ibge_cod}{Numérico ou caractere. Código ibge do(s) município(s) brasileiro(s)}
}
\description{
Esta função extrai os valores compilados pelo portal Brasil.io, que recolhe boletins informativos e casos do coronavírus por minicípio e por dia (disponível em: https://brasil.io/dataset/covid19/caso). A função salva o resultado no disco.
}
