% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corona_minsaude.R
\name{get_corona_minsaude}
\alias{get_corona_minsaude}
\title{Extrai dados de COVID-19 para o Brasil da página do Ministério da Saúde}
\usage{
get_corona_minsaude(dir = "output", filename = "minsaude", uf = NULL)
}
\arguments{
\item{dir}{Diretório onde salvar o arquivo}

\item{filename}{Nome do arquivo, valor predeterminado "minsaude"}

\item{uf}{Caractere indicando a abreviação do(s) estado(s) brasileiro(s)}
}
\description{
Esta função extrai os valores compilados pelo Ministério da Saúde do Brasil
(disponíveis em: 'https://covid.saude.gov.br/') e salva o resultado no disco.
}
