% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_srag_minsaude.R
\name{get_srag_minsaude}
\alias{get_srag_minsaude}
\title{Extrai dados de SRAG para o Brasil da página do Ministério da Saúde}
\usage{
get_srag_minsaude(dir = "outputs", filename = "minsaude_srag", uf = NULL)
}
\arguments{
\item{dir}{Diretório onde salvar o arquivo, valor predeterminado "outputs"}

\item{filename}{Nome do arquivo, valor predeterminado "minsaude_srag"}

\item{uf}{Caractere indicando a abreviação do(s) estado(s) brasileiro(s)}
}
\description{
Esta função extrai os valores referentes à Síndrome Respiratória Aguda Grave
compilados pelo Ministério da Saúde do Brasil (disponíveis em: 'https://covid.saude.gov.br/')
e salva o resultado no disco.
}
