% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif_corona_br.R
\name{gif_corona_br}
\alias{gif_corona_br}
\title{Mapa do número de casos por estado}
\usage{
gif_corona_br(df, prop_pop = TRUE)
}
\arguments{
\item{df}{Data frame contendo o resultado da busca de `get_corona_br()`}

\item{prop_pop}{Lógico. Exibir gráfico com número de casos proporcional à população? Padrão prop_pop = TRUE}
}
\description{
Esta função faz o mapa do número de casos por estado brasileiro.
}
