% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_corona_br.R
\name{map_corona_br}
\alias{map_corona_br}
\title{Mapa do número de casos por estado}
\usage{
map_corona_br(df, prop_pop = TRUE)
}
\arguments{
\item{df}{Data frame formatado com a função `format_corona_br()`}

\item{prop_pop}{Lógico. Exibir gráfico com número de casos proporcional à população? Padrão prop_pop = TRUE}
}
\description{
Esta função faz o mapa do número de casos por estado brasileiro.
}
