% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uf.R
\name{plot_uf}
\alias{plot_uf}
\title{Plot o número de casos (confirmados ou óbitos) por Estados com dados do Brasil.io}
\usage{
plot_uf(df, prop_pop = TRUE, tipo = "casos", n = 10)
}
\arguments{
\item{df}{Data frame por estado formatado com a função `format_corona_br()`}

\item{prop_pop}{Lógico. Exibir gráfico com número de casos proporcional à população? Padrão prop_pop = TRUE}

\item{tipo}{Variável a ser exibida no eixo y. Padrão `tipo = "casos"` para mostrar os casos confirmados. Use `tipo = "obitos"` para o gráfico com o número de óbitos}

\item{n}{Inteiro. Número mínimo de estados para o gráfico}
}
\value{
objeto ggplot
}
\description{
Plot o número de casos (confirmados ou óbitos) por Estados com dados do Brasil.io
}
\examples{
\dontrun{
dados <- get_corona_br(by_uf = TRUE)
dados_format <- format_corona_br(dados)
plot_uf(df = dados_format)
}
}
