% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uf2.R
\name{plot_uf2}
\alias{plot_uf2}
\title{Plot o número de casos (confirmados ou óbitos) por Estados com dados do Brasil.io}
\usage{
plot_uf2(df, prop_pop = TRUE, tipo = "casos", n_casos = 1500, n_obitos = 100)
}
\arguments{
\item{df}{Data frame por estado formatado com a função `format_corona_br()`}

\item{prop_pop}{Lógico. Exibir gráfico com número de casos proporcional à população? Padrão prop_pop = TRUE}

\item{tipo}{Variável a ser exibida no eixo y. Padrão `tipo = "casos"` para mostrar os casos confirmados. Use `tipo = "obitos"` para o gráfico com o número de óbitos}

\item{n}{Inteiro. Número mínimo de casos para comparação entre estados}

\item{dir}{Caractere. Nome do diretório onde salvar a animação. Usar apenas se anim = TRUE.}
}
\value{
objeto ggplot
}
\description{
Plot o número de casos (confirmados ou óbitos) por Estados com dados do Brasil.io
}
