% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corona.R
\name{get_corona}
\alias{get_corona}
\title{Extrai dados de corona vírus para o Brasil}
\usage{
get_corona(dir = "output/", filename = "corona_brasil")
}
\arguments{
\item{dir}{diretório onde salvar o arquivo}

\item{filename}{nome do arquivo}
}
\description{
Esta função extrai os valores compilados pelo ministério da Saúde do Brasil (disponível em: 'http://plataforma.saude.gov.br/novocoronavirus/#COVID-19-brazil') e salva o resultado no disco.
}
