% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corona_br.R
\name{get_corona_br}
\alias{get_corona_br}
\title{Extrai dados de coronavírus para o Brasil de Brasil.io}
\usage{
get_corona_br(
  dir = "output",
  filename = "corona_brasil",
  cidade = NULL,
  uf = NULL,
  ibge_cod = NULL,
  by_uf = FALSE
)
}
\arguments{
\item{dir}{Diretório onde salvar o arquivo}

\item{filename}{Nome do arquivo, valor predeterminado "minsaude"}

\item{cidade}{Caractere indicando o(s) nome(s) do(s) município(s) brasileiro(s). Atenção, fornecer também o estado ou o código ibge para evitar ambiguidade. "Importados" extrai os casos importados, se disponíveis.}

\item{uf}{Caractere indicando a abreviação do(s) estado(s) brasileiro(s)}

\item{ibge_cod}{Numérico ou caractere. Código ibge do(s) município(s) ou estado(s) brasileiro(s). O código dos municípios tem 7 cifras e o código dos estados 2}

\item{by_uf}{Lógico. Padrão by_uf = FALSE. Usar by_uf = TRUE se quiser os dados apenas por UF independente do município. Usar apenas quando não fornecer `cidade` ou `ibge_cod` de um município}
}
\description{
Esta função extrai os valores compilados pelo portal Brasil.io, que recolhe boletins informativos e casos do coronavírus por minicípio e por dia (disponível em: https://brasil.io/dataset/covid19/caso). A função salva o resultado no disco e escreve um arquivo com os metadados da requisição (metadado_corona_br.csv).
}
