% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_uf.R
\name{plot_uf}
\alias{plot_uf}
\title{Plot o número de casos (confirmados ou mortes) por Estados com dados do Brasil.io}
\usage{
plot_uf(
  df,
  prop_pop = TRUE,
  tipo = "casos",
  n = 100,
  anim = FALSE,
  dir = "figs"
)
}
\arguments{
\item{df}{Data frame por estado formatado com a função `format_corona_br()`}

\item{prop_pop}{Lógico. Exibir gráfico com número de casos proporcional à população? Padrão prop_pop = TRUE}

\item{tipo}{Variável a ser exibida no eixo y. Padrão `tipo = "casos"` para mostrar os casos confirmados. Use `tipo = mortes` para o gráfico com o número de mortes}

\item{n}{Inteiro. Número mínimo de casos para comparação entre estados}

\item{anim}{Lógico. Retornar animação com evolução do número de casos ao longo do tempo. Padrão anim = FALSE}

\item{dir}{Caractere. Nome do diretório onde salvar a animação. Usar apenas se anim = TRUE.}
}
\value{
objeto ggplot
}
\description{
Plot o número de casos (confirmados ou mortes) por Estados com dados do Brasil.io
}
\examples{
dados <- get_corona_br(by_uf = TRUE)
dados_format <- format_corona_br(dados)
plot_uf(df = dados)

}
